+++ app/Fields/Uitype/Password.php
@@ -111,21 +111,4 @@ class Password extends Text implements Uitype
             $field->data->repeated = true;
         }
     }
-
-    /**
-     * Returns some specific options relative to the field, used by uccello/module-designer
-     *
-     * @return array
-     */
-    public function moduleDesignerConfig()
-    {
-        return [
-            [
-                'label' => trans('uccello::config.uitypes.global.repeat_field'),
-                'type' => 'boolean',
-                'attribute' => 'repeated',
-                'default' => true,
-            ],
-        ];
-    }
 }