+++ app/Models/Group.php
@@ -2,14 +2,14 @@
 
 namespace Uccello\Core\Models;
 
+use App\Models\UccelloModel;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Spatie\Searchable\Searchable;
 use Spatie\Searchable\SearchResult;
-use Uccello\Core\Database\Eloquent\Model;
 use Uccello\Core\Support\Traits\UccelloModule;
-use App\User;
+use App\Models\User;
 
-class Group extends Model implements Searchable
+class Group extends UccelloModel implements Searchable
 {
     use SoftDeletes;
     use UccelloModule;
@@ -47,7 +47,7 @@ class Group extends Model implements Searchable
     ];
 
     public function getSearchResult(): SearchResult
-    {    
+    {
         return new SearchResult(
             $this,
             $this->recordLabel