+++ database/migrations/2018_04_15_000011_create_blocks_table.php
@@ -25,8 +25,8 @@ class CreateBlocksTable extends Migration
 
             // Foreign keys
             $table->foreign('module_id')
-                ->references('id')->on($this->tablePrefix.'modules')
-                ->onDelete('cascade');
+                    ->references('id')->on($this->tablePrefix.'modules')
+                    ->onDelete('cascade');
 
             $table->foreign('tab_id')
                     ->references('id')->on($this->tablePrefix.'tabs')