+++ database/migrations/2018_04_15_000015_create_filters_table.php
@@ -18,19 +18,12 @@ class CreateFiltersTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('module_id');
             $table->unsignedInteger('domain_id')->nullable();
-
-            // Compatibility with Laravel < 5.8
-            if (DB::getSchemaBuilder()->getColumnType('users', 'id') === 'bigint') { // Laravel >= 5.8
-                $table->unsignedBigInteger('user_id')->nullable();
-            } else { // Laravel < 5.8
-                $table->unsignedInteger('user_id')->nullable();
-            }
-
+            $table->unsignedBigInteger('user_id')->nullable();
             $table->string('name');
             $table->string('type');
             $table->text('columns');
             $table->text('conditions')->nullable();
-            $table->string('order_by')->nullable();
+            $table->string('order')->nullable();
             $table->boolean('is_default')->default(false);
             $table->boolean('is_public')->default(false);
             $table->text('data')->nullable();