+++ resources/views/modules/default/detail/main.blade.php
@@ -36,9 +36,7 @@
 {{-- Custom links --}}
 @section('top-action-buttons')
     <div class="action-buttons right-align">
-    @section ('custom-links')
-        @include('uccello::modules.default.detail.links')
-    @show
+    @yield('custom-links')
     </div>
 @append