+++ resources/views/modules/default/list/modal/export.blade.php
@@ -1,9 +1,9 @@
-<div id="exportModal" class="modal">
+<div id="exportModal" class="modal modal-fixed-footer">
     <div class="modal-content">
-        <h4>
+        <h5>
             <i class="material-icons left primary-text">cloud_download</i>
             {{ uctrans('modal.export.title', $module) }}
-        </h4>
+        </h5>
 
         <p>{{ uctrans('modal.export.description', $module) }}</p>
 
@@ -11,12 +11,15 @@
             <div class="col s12">
                 <div class="input-field">
                     <select id="export_format">
-                        <option value="xlsx">{{ uctrans('modal.export.format.xlsx', $module) }}</option>
-                        <option value="xls">{{ uctrans('modal.export.format.xls', $module) }}</option>
-                        <option value="ods">{{ uctrans('modal.export.format.ods', $module) }}</option>
-                        <option value="csv">{{ uctrans('modal.export.format.csv', $module) }}</option>
-                        <option value="html">{{ uctrans('modal.export.format.html', $module) }}</option>
-                        <option value="pdf">{{ uctrans('modal.export.format.pdf', $module) }}</option>
+                        @section('export-formats')
+                            <option value="xlsx">{{ uctrans('modal.export.format.xlsx', $module) }}</option>
+                            <option value="xls">{{ uctrans('modal.export.format.xls', $module) }}</option>
+                            <option value="ods">{{ uctrans('modal.export.format.ods', $module) }}</option>
+                            <option value="csv">{{ uctrans('modal.export.format.csv', $module) }}</option>
+                            <option value="html">{{ uctrans('modal.export.format.html', $module) }}</option>
+                        @show
+
+                        @yield('extra-export-formats')
                     </select>
                     <label>{{ uctrans('modal.export.format_label', $module) }}</label>
                 </div>
@@ -56,12 +59,28 @@
                     </label>
                 </p>
 
+                @yield('extra-export-options')
+
                 <input type="hidden" id="export_with_descendants" @if ($seeDescendants) value="1" @else value="0" @endif />
             </div>
+
+            @yield('extra-export-modal-content')
         </div>
     </div>
+
+    @section('export-modal-footer')
     <div class="modal-footer">
+        @yield('before-cancel-export-button')
+
+        {{-- Cancel button --}}
         <a href="javascript:void(0)" class="btn-flat waves-effect modal-close">{{ uctrans('button.cancel', $module) }}</a>
+
+        @yield('before-export-button')
+
+        {{-- Export button --}}
         <a href="javascript:void(0)" class="btn-flat waves-effect modal-close green-text export">{{ uctrans('button.export', $module) }}</a>
+
+        @yield('after-export-button')
     </div>
+    @show
 </div>